Update Si5351_VFO_0v2.ino
Declared the variable "Turns" as an int rather than a double

+++ Si5351_VFO_0v2.ino
@@ -178,7 +178,7 @@ int BannerPos=0;
 int c1=0;
 int c2=0;
 int c3=0; 
-double Turns = 0;
+int Turns = 0;   // was formerly declared as a double
 
 int Menu = 0;
